Break out items rendering to external script
